+++ src/main/java/com/codetaylor/mc/pyrotech/modules/hunting/init/PacketInitializer.java
@@ -3,6 +3,7 @@ package com.codetaylor.mc.pyrotech.modules.hunting.init;
 import com.codetaylor.mc.athenaeum.network.IPacketRegistry;
 import com.codetaylor.mc.pyrotech.modules.hunting.network.CSPacketCapabilitySyncSpearRequest;
 import com.codetaylor.mc.pyrotech.modules.hunting.network.SCPacketCapabilitySyncSpear;
+import com.codetaylor.mc.pyrotech.modules.hunting.network.SCPacketParticleMud;
 import net.minecraftforge.fml.relauncher.Side;
 
 public final class PacketInitializer {
@@ -12,6 +13,12 @@ public final class PacketInitializer {
     registry.register(
         SCPacketCapabilitySyncSpear.class,
         SCPacketCapabilitySyncSpear.class,
+        Side.CLIENT
+    );
+
+    registry.register(
+        SCPacketParticleMud.class,
+        SCPacketParticleMud.class,
         Side.CLIENT
     );